Refactor pbench namespace
Split out server bits and agent bits into seperate namespaces
so its easier to read and makes more sense.
